Game-Recommander/requirements.txt
neo4j
python-dotenv
pandas
numpy
langchain
langchain-openai
langchain-community
langchain-neo4j
langchain-groq
openai
streamlit
nltk
spacy
scikit-learn
tqdm